pyproject.toml
[tool.poetry]
name = "istar"
version = "0.1.0"
description = ""
authors = ["Daiwei Zhang <daviddaiweizhang@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.25.2"
pillow = "^10.0.0"
pandas = "^2.1.0"
scikit-image = "^0.21.0"
opencv-python = "^4.8.0.76"
einops = "^0.6.1"
torch = {version = "^2.0.1+cu118", source = "pytorch-gpu-src"}
torchvision = {version = "^0.15.2+cu118", source = "pytorch-gpu-src"}
tomli = "^2.0.1"
pytorch-lightning = "^2.0.8"
matplotlib = "^3.7.2"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
ipython = "^8.15.0"
black = "^23.7.0"
tensorboard = "^2.14.0"

[tool.poetry.scripts]
istar = "istar.__main__:cli"

[[tool.poetry.source]]
name = "pytorch-gpu-src"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
